Basic loop with SDL event handler just for window to display properly
