+++ src/platform/input.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <SDL3/SDL.h>
+
+namespace platform {
+
+bool eventHandler(SDL_Event* event);
+
+} //namespace platform
