+++ src/platform/platform.hpp
@@ -1,23 +1,6 @@
 #pragma once 
 
 #include <SDL3/SDL.h>
-#include "platform/input.hpp"
+#include <platform/input.hpp>
+#include <platform/window.hpp>
 
-namespace platform {
-
-struct windowArg {
-	const char *title;
-	int width;
-	int height;
-};
-
-class window {
-public:
-	void windowInit(windowArg args);
-	void windowClear();
-
-private:
-	SDL_Window *window;
-};
-
-} //namespace platform
